added confirmation type, fixed linter issue

+++ src/main/java/com/podzilla/order/dtos/LocationDTO.java
@@ -4,12 +4,11 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class LocationDTO {
-    private BigDecimal first;  // Latitude
-    private BigDecimal second; // Longitude
+    private double first;  // Latitude
+    private double second; // Longitude
 }
